requirements.txt
python-telegram-bot
requests
yandex-translater